requirements.txt
asgiref==3.6.0
astroid==2.15.4
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
dicebear==2.6.1
dill==0.3.6
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-js-asset==2.0.0
django-lifecycle==1.0.0
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.4
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
urlman==2.0.1
wrapt==1.15.0
